refactor: rename TaskItem → TaskCard; add feature props for reuse across views
